Quick Expense parameter description improvements
These edits 1) provide a link to the list of Spend Category Codes (which are frustratingly not found in the API Reference section but instead in the Reference section under Tools & Support ... why??), 2) clarify that the mileage-related Spend Category Codes simply don't work for Quick Expense, and 3) remove the reference to the Get Expense Group Configuration function which isn't available to App for Me-class integrations and is superfluous. Erik Hansen should review my edits.