Change from 'Authentication' to 'Authorization'
Callers of APIs almost never authenticate directly  -- I can't think of a scenario EXCEPT for internal service usage, and for that the authentication is X.509 certificates, not an Authentication API). External callers always get Authorization to use an API on behalf of another, and there may be authentication involved in that process.