Update expense-report-get.markdown
This API is not officially deprecated and it cannot be until v4 is released.  note: the v3 Entries API should not be used due to a major issue with the response related to reports being created by a Delegate - the report content does not come back in the response.  So, when we refer devs to this page, they instantly ask about the deprecation reference.  Can this be moved to a page where it doesn't state "deprecated" ?