Update apidoc.markdown
can you also please bold the following text and use a different color such as blue?
"When obtaining the token, the token's geolocation will be included in the response. The token's geolocation should be stored along with the token. The Developer's app will then be able to make subsequent calls using the correct end points per GEO location."