Update fetch-attendee.markdown
This missing information occasionally results in unnecessary tickets from clients. E.g. salesforce 14406523 or JIRA CRMC-141157. 
It is the very first time I propose a change and I don't know if this process is good. Would you mind letting me know if this is ok on Slack or Skype or email (Pavel Kuchyna) ?