Update post-list-search-request.markdown
The  fields that were added as part of this Documentation update, are being sent to the designated Client Endpoint from the Concur Callout Server. Currently they are not documented and this has caused confusion to the customers Implementing this Callout. An example is the  CRMC-126622 created for Siemens. The proposed changed will mitigate the gap between documentation and what is being sent to the Customers as part of the XML.
The fields description is not known to me  and this is the reason that an empty space has been used in this part of the Documentation. 

The second change is regarding the Required Fields of the Response that is being sent as part of the Callout and they should be marked as Required. The reason is that the lack of marking them in the documentation creates the impression to the customers that they can skip sending them back which will cause failure to the Callout functionality. This has created problems in Callout Client Implementation Projects like Siemens, Generalli and others in the past, by following the existing Documentation.